requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
boto3==1.37.25
botocore==1.37.25
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
Events==0.5
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-aws==0.2.19
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.30
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
mysql-connector-python==9.3.0
numpy==2.2.4
opensearch-py==2.8.0
orjson==3.10.16
packaging==24.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyMuPDF==1.25.5
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.19.0
zstandard==0.23.0
